♻️ refactor(repo-search): update state management
