+++ packages/client/src/features/repo-search/ui/RepoSearchForm.tsx
@@ -1,27 +1,29 @@
+import type { FormEventHandler } from "react";
 import { TextInput } from "../../../shared";
-import { useRepoUrlForm } from "../model/useRepoUrlForm";
-import { useCommits } from "../model/useCommits";
 
-export function RepoSearchForm() {
-  const { repoUrl, setRepoUrl, submittedRepoUrl, onSubmitUrl } =
-    useRepoUrlForm();
-  const commitsQuery = useCommits(submittedRepoUrl);
+export interface RepoSearchFormProps {
+  value: string;
+  onChange: (value: string) => void;
+  onSubmit: FormEventHandler<HTMLFormElement>;
+  error?: string;
+}
 
-  const inputErrorMessage =
-    commitsQuery.error instanceof Error
-      ? commitsQuery.error.message
-      : undefined;
-
+export function RepoSearchForm({
+  value,
+  onChange,
+  onSubmit,
+  error,
+}: RepoSearchFormProps) {
   return (
-    <form className="flex w-full gap-3 items-center" onSubmit={onSubmitUrl}>
+    <form className="flex w-full gap-3 items-center" onSubmit={onSubmit}>
       <TextInput
         label="Repository URL"
         placeholder="https://github.com/owner/repository"
-        value={repoUrl}
-        onChange={(event) => setRepoUrl(event.target.value)}
+        value={value}
+        onChange={(event) => onChange(event.target.value)}
         fullWidth
         required
-        errorText={inputErrorMessage}
+        errorText={error}
       />
     </form>
   );
