+++ packages/client/src/pages/ui/HomePage.tsx
@@ -1,10 +1,10 @@
-import { RepoSearchForm } from "../../features/repo-search/ui/RepoSearchForm";
+import { CommitBrowser } from "../../widgets/commit-browser/CommitBrowser";
 
 export function HomePage() {
   return (
     <div className="space-y-6">
       <h1 className="text-3xl font-bold text-gray-900">Home</h1>
-      <RepoSearchForm />
+      <CommitBrowser />
     </div>
   );
 }
