+++ packages/client/src/widgets/commit-browser/CommitBrowser.tsx
@@ -1,0 +1,24 @@
+import { useRepoUrlForm } from "../../features/repo-search/model/useRepoUrlForm";
+import { useCommits } from "../../features/repo-search/model/useCommits";
+
+import { RepoSearchForm } from "../../features/repo-search/ui/RepoSearchForm";
+
+export function CommitBrowser() {
+  const { repoUrl, setRepoUrl, submittedRepoUrl, onSubmitUrl } =
+    useRepoUrlForm();
+  const commitsQuery = useCommits(submittedRepoUrl);
+  return (
+    <div>
+      <RepoSearchForm
+        value={repoUrl}
+        onChange={setRepoUrl}
+        onSubmit={onSubmitUrl}
+        error={
+          commitsQuery.error instanceof Error
+            ? commitsQuery.error.message
+            : undefined
+        }
+      />
+    </div>
+  );
+}
